split PaymentPage into components and separate styles
refactor(payment): modularize PaymentPage
- removed old PaymentPage.css
- added components (AddCardForm, BookingDetails, CardSelection, etc.)
- added related styles, utils, and hook
