+++ src/assets/styles/payment/BookingDetails.css
@@ -1,0 +1,149 @@
+.payment-card {
+  background: white;
+  border-radius: 1rem;
+  padding: 2rem;
+  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
+  border: 1px solid #e5e7eb;
+}
+
+.card-title {
+  font-size: 1.5rem;
+  font-weight: 700;
+  color: #111827;
+  margin-bottom: 1.5rem;
+}
+
+.station-info {
+  margin-bottom: 2rem;
+}
+
+.station-name {
+  font-size: 1.25rem;
+  font-weight: 600;
+  color: #111827;
+  margin-bottom: 0.5rem;
+}
+
+.station-address {
+  color: #6b7280;
+  font-size: 0.875rem;
+  margin: 0;
+}
+
+.details-grid {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 1rem;
+  margin-bottom: 2rem;
+}
+
+.detail-item {
+  padding: 1rem;
+  background-color: #f9fafb;
+  border-radius: 0.5rem;
+  border: 1px solid #e5e7eb;
+}
+
+.detail-header {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  margin-bottom: 0.5rem;
+}
+
+.detail-icon {
+  font-size: 1.25rem;
+}
+
+.detail-label {
+  font-size: 0.75rem;
+  font-weight: 600;
+  color: #6b7280;
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+}
+
+.detail-value {
+  color: #111827;
+  font-weight: 500;
+  margin: 0;
+  font-size: 0.875rem;
+}
+
+.cost-section {
+  border-top: 1px solid #e5e7eb;
+  padding-top: 1.5rem;
+}
+
+.cost-title {
+  font-size: 1.125rem;
+  font-weight: 600;
+  color: #111827;
+  margin-bottom: 1rem;
+}
+
+.cost-list {
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+
+.cost-item {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0.5rem 0;
+}
+
+.cost-label {
+  color: #6b7280;
+  font-size: 0.875rem;
+}
+
+.cost-value {
+  color: #111827;
+  font-weight: 500;
+  font-size: 0.875rem;
+}
+
+.cost-total {
+  border-top: 2px solid #e5e7eb;
+  padding-top: 1rem;
+  margin-top: 1rem;
+}
+
+.cost-total .cost-item {
+  font-size: 1.125rem;
+  font-weight: 700;
+}
+
+.cost-total .cost-value {
+  color: #059669;
+}
+
+.cost-item.surcharge {
+  background-color: #fef3c7;
+  padding: 0.75rem;
+  border-radius: 0.5rem;
+  border: 1px solid #fed7aa;
+}
+
+.cost-item.surcharge .cost-label {
+  color: #d97706;
+  font-weight: 600;
+}
+
+.cost-item.surcharge .cost-value {
+  color: #d97706;
+  font-weight: 600;
+}
+
+@media (max-width: 640px) {
+  .details-grid {
+    grid-template-columns: 1fr;
+  }
+  
+  .payment-card {
+    padding: 1.5rem;
+  }
+}
