+++ src/assets/styles/payment/PaymentHeader.css
@@ -1,0 +1,74 @@
+.payment-header {
+  background: white;
+  color: #10b981;
+  padding: 1.5rem 0;
+  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
+  border-bottom: 2px solid #10b981;
+  margin-top: 80px;
+  position: relative;
+  z-index: 10;
+}
+
+.header-content {
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 0 1rem;
+}
+
+.logo-section {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.logo-icon {
+  width: 4rem;
+  height: 4rem;
+  background-color: #10b981;
+  color: white;
+  border-radius: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 2rem;
+  transition: all 0.2s ease-in-out;
+  box-shadow: 0 4px 12px rgba(16, 185, 129, 0.3);
+}
+
+.logo-icon svg {
+  width: 2rem;
+  height: 2rem;
+}
+
+.logo-icon:hover {
+  transform: scale(1.05);
+  background-color: #059669;
+}
+
+.logo-text {
+  font-size: 1.75rem;
+  font-weight: 700;
+  margin: 0;
+  color: #10b981;
+}
+
+@media (max-width: 768px) {
+  .logo-icon {
+    width: 3rem;
+    height: 3rem;
+    font-size: 1.5rem;
+  }
+  
+  .logo-icon svg {
+    width: 1.5rem;
+    height: 1.5rem;
+  }
+  
+  .logo-text {
+    font-size: 1.5rem;
+  }
+  
+  .payment-header {
+    padding: 1rem 0;
+  }
+}
